starlark: move pip_install() from distribution to executable
Since some standalone Python distributions can now support linking libpython
statically and dynamically, we need packaging functionality to be aware of the
environment it is targeting. Our choices are to add arguments to the existing
methods on distribution types specifying the settings or to move methods to the
binary builders, which already have the necessary context. We choose the latter.