python-packaging: define new methods taking a location
There are a lot of different APIs that only vary slightly in
semantic behavior. I think it makes sense to consolidate the APIs
by having them accept a location to add a resource to. Among other
things, this will help enforce consistency and make it easy
to parameterize behavior since we won't have to handle differing
function dispatch.