python-packaging: use rich type to capture library dependencies
This is needed for parity with distribution extensions. I suspect we'll
also want this for when we need to record richer metadata needed for
linking. It is quite apparent that linking will likely fail in a lot
of scenarios due to information loss currently in the code. But we can
deal with that another time.