py_packaging: use LibraryDependency instead of LibraryDepends
The new type in python-packaging should be functionally equivalent,
rendering this type unnecessary.

The more I touch this code the more I realize how fragile it is.
Hopefully once we get closer to a final state we can start to
implement some tests for all the functionality, so I'm worried there
are a lot of subtle bugs lingering in everything related to
extensions and library handling.