py_packaging: pass PythonExtensionModule into add_builtin_distribution_extension_module
PythonExtensionModule now has enough feature parity with the
distribution extension module that we can start to replace the latter
with the former. Let's start on our embedded resources type, which
is the lowest level API.

We might be able to delete some of the methods once everything is fully
ported over since there are similar methods only varying in the type
they accept. We'll cross this bridge later.