py_packaging: remove DistributionExtensionModule
It took a few hours of work, but we are now fully consolidated onto
PythonExtensionModule and no longer have multiple primitives in the
core packaging code for representing extension modules!

There's still a bit of clean-up with regards to function names
referencing "distribution" extensions. But we're already much, much
better than we were previously.