starlark: remove filtering from PythonDistribution.extension_modules()
Filtering belongs at the binary building layer. I think it makes
sense to expose all extensions on this API. Plus the new code is simpler
in that we don't have to manage filtering/policies in more than 1
place.

As part of this, we expanded the PythonDistribution trait to emit
all extensions in the distribution.