py_packaging: manually handle symlinks on Windows
We want to make it possible to utilize distributions from any platform
on any platform to better facilitate testing. But non-Windows
distributions may have symlinks in tar archives and the tar crate
will attempt to blindly create symlinks, even if they aren't supported.

This commit teaches the tar extraction code to handle symlinks on
Windows. When a symlink is encountered, it will copy the linked file
into what would normally be the symlink.