py_packaging: extract standalone builder to own module
standalone_distribution.rs was getting a bit too large for my taste.
This commit splits out the binary builder part to its own module.

Some fields of the standalone distribution had to be made public
to facilitate access outside the module. And logic for
constructing a standalone builder had to be moved to the new
module to avoid having to make fields public.