python-packaging: add tests for PrePackagedResource -> Resource conversion
We verify that every field is converted. The tests are rather verbose
and add a lot of lines. But this conversion is important to have test
coverage for.

As part of this, we implement a fake bytecode compiler to facilitate
testing. This was made possible by recent commits to define a trait
for the bytecode compiler.