python-packaging: add test coverage for to_prepared_python_resources()
We expand existing tests adding resources to call
to_prepared_python_resources() and verify that the prepared resources
match expectations. Test coverage is not yet comprehensive. But it is
better than what existed before. The reasons I stopped here is because
the next test function is buggy and I want to fix that as I add test
coverage.