python-packaging: copy all resources entries
Without this, we fail to copy over entries for extensions and other
resource types which should be preserved as part of packaging.

As part of this commit, we remove the explicit prefix storing in the
pre packaged resource because it is already incorporated into the path
stored internally and we were double prefixing the install path, which
the added test coverage helped us uncover.