python-packaging: install shared library dependencies next to extension
If we're installing shared library dependencies for an extension, we
want them placed next to the extension file and not directly in the
prefix.

I'm not 100% confident this is the correct behavior on all platforms.
But it intuitively feels correct. We'll probably want to verify this
with some real world extensions.