py_packaging: add a type representing libpython linking context
Currently, state influencing libpython generation lives in a number
of different types. I'd like to unify the tracking of this state in
a single type.

This commit establishes a new `LinkingContext` type to eventually
track all this state. We start by capturing a list of object files
to link. It isn't populated but the object files field is used during
linking. Future commits will implement functionality.