py_packaging: add distribution's core object files to linking context
This effectively moves population of the object files from
link_libpython() to the binary builder.

The old code copied the object file into a temp directory. Why, I'm
not sure. We could potentially always copy file references present
in the linking context. I suppose we'll do that if there is any issue
with the change.