py_packaging: establish method for adding a PythonExtensionModule
This currently makes a lot of assumptions. But it is a start.

As part of this, we add test coverage for adding distribution
extensions using this API.

Some of the behavior is not correct. Our goal is to establish test
coverage first so we can observe changes in behavior as the code
changes.