py_packaging: only return builder from new_builder()
Now that we return the actual type instead of a trait, we have
access to private fields, which includes access to the distribution.

In a few places, we were using a reference to an internal object,
which confused the borrow checker. So we had to clone some objects
as a hacky workaround.