py_packaging: add extension module variants with object files
We want to verify how extension modules with just object files or a
mix of object files and a shared library are handled. So add test
coverage.

Our new API rejects these extensions since they aren't part of the
standard library. We'll update the tests once the new code is
implemented.