py_packaging: remove remaining add_*extension one-offs
add_python_extension_module() is our new unified API. Let's use it
everywhere.

I initially set out to remove just the in-memory add variation. However,
once I did this, I realized that the generic remaining method always
called add_python_extension_module() and the only logic was determining
what to pass as the 2nd argument. But add_python_extension_module()
already has logic for determining the appropriate resource location
given the policy, so this function was reduced to a proxy always calling
None. So I decided to remove it in the same commit.

As part of this change, behavior in the Linux and macOS tests for
`libpython_link_mode: BinaryLibpythonLinkMode::Dynamic` started failing
for unknown reasons. Those tests seemed suspect to me, so I deleted them
for the moment.

The change in test_windows_dynamic_extensions_sanity is a bit more
interesting and points to a legitimate regression. Looking at this
in more detail, the old logic was a bit wonky in that it was using the
libpython link mode to determine resource location. This didn't make
sense. So the change to loading extensions as built-ins does seem to
make sense. However, I do think a case can be made that stdlib
extensions shipped as shared libraries should be preserved that way
if the resources policy allows it. But we don't have test coverage of
default distribution extension behavior given different resource
policies. We should probably do that.