py_packaging: add libpython link mode variations to tests
These tests were dropped in the last commit. Let's add them back in.

The link mode doesn't currently seem to affect extension module
placement, which I think makes sense, as the choice of whether to
make an extension built-in versus a shared library is more about the
resources policy and not whether libpython will be statically or
dynamically linked into the executable.