starlark: obtain an EmbeddedPythonConfig() by calling a Rust API
This is simpler than going through Starlark dispatch. To ensure
behavior is consistent with the new Starlark API, we add a test
that compares the config instances.

There's probably room to consolidate logic further and to e.g. use
None as the default value for arguments. I've added a TODO to track
this.