starlark: port to Environment -> TypeValues upstream change
The upstream crate stopped given Starlark functions access to the
full Environment and only gives them access to TypeValues. This means
we could no longer access variables or global functions from the context
of a Rust-implemented Starlark function.

Our (somewhat hacky) workaround mainly deployed in previous commits was
to create a dummy `PyOxidizerContext` type, register it as a type
value, then alias various global variables to attributes on this type
value. This facilitates access to global functions and variables from
the context of a type value. This is arguably cheating. But it does
allow us to circumvent limitations of the new API and continue with
our port to version 0.3 of the crate.

This compatibility issue was discussed at
https://github.com/google/starlark-rust/issues/265. Although I dropped
the ball a bit and failed to follow up with the upstream crate
maintainer. Since I have no devised a workaround, I may continue to let
that issue linger. Although I would like to see Starlark Rust functions
regain the ability to access global variables and functions from their
called context. That context is available to Python functions: I think
it should be available to Rust code as well.