starlark: pass None as default argument values to PythonInterpreterConfig
If None is passed, we now take the default value from the default
instance. This prevents a potential for the default instance and the
argument default values to be out of sync.

As part of this, I discovered a bug around handling of
`terminfo_resolution`. The docs said to pass the value `'none'` to
indicate no resolution. However, we failed to handle this value at
run-time! We added test coverage for the documented behavior.

There is backwards incompatible behavior when passing
`terminfo_resolution=None`: before it would set the `None`
mode. Now, it sets the default, which is `Dynamic`.