starlark: consolidate Python module source methods into one accepting location
I want to overhaul the Starlark APIs for handling locations. We don't
need separate API flavors for each location. Instead, we can pass a
location as an optional argument.

This will eventually lead to Starlark having a similar API as the
internal binary builder trait. This will be a backwards incompatible
user transition. But the end state will be much simpler. And it will
be easier to add more power to the location routing, as the policy
gains more complexity in the future.