starlark: start to define context for how resources are added
As part of improving how resources interact with the resource
collector, we want to move the processing of the packaging policy
from resource add time to when that resource is created. Our end goal
is for the Starlark resource types to all define context for how
they will be added to the resource collector. When the resource
is created, the policy will be queried and this context will be
populated. This will enable Starlark code to query and mutate the
collector interaction via attributes on the resource instead of
passing arguments to add_* methods.

This commit begins that process by defining a new struct to hold
this state and adding that context to the Python module source
Starlark type.

I'm not yet convinced we aren't going to have to refactor internals
substantially as part of implementing this change. We don't yet
use this state to influence resource adding. But this seemed like
a logical stopping point, as the patch was already getting quite long.