starlark: define a trait for add collection context
Multiple resource types are going to be exposing attributes for
the add collection context. The logic for attribute interaction
can be shared between the types. So let's establish a trait for
that logic and use it on `PythonModuleSourceValue`, which is the
only type currently exposing any functionality.

We allow the add collection context to be None, as I think it makes
sense for it to be optional, as we can create resources without
a resources policy available. (e.g. obtaining resources in a
`PythonDistribution` instance.) We will still need to think through
how we want objects to behave. But with the logic being in a trait,
it should make things much easier to refactor.