python-packaging: use a single enum representation for extension modules
With a unified `PythonExtensionModule` type that can represent both
static and dynamic extension modules, we no longer need multiple
enum variants for extension modules.

We may have subtly broken the hacked distutils building mode
with this change. If we did, that's our fault for lacking test
coverage for this scenario. The plan is to shore up test coverage
in this area before the next release. But first, we want to get
the packaging policy code in a better state. And that requires
improving resource representation.