python-packaging: store a Cow<T> inside PythonResource
This allows us to have PythonResource representing a borrowed
value, which could be useful.

This required annotating a lot of lifetimes. But the compiler is
happy and the implementation should be sound.

As part of this, I converted a bunch of manual enum variant
constructions to use `.into()`, which keeps things more concise.
We may have thrown away some type conversion checking. But I think
garbage-in-garbage-out applies and we can rely on From<T> being
implemented properly.

We could potentially implement `AsRef<PythonResource>` now. But
we'll save that for another commit.