starlark: method to apply packaging policy to a resource
This method will enable us to define or reset the collection add
context on a resource, which will be useful when "moving" resources
from a distribution context to a binary building one.

Now that `PythonResource` is implemented in terms of `Cow`, we
are able to convert a ref to a Python resource to a `PythonResource`
with minimal effort.