starlark: define a PythonPackagingPolicy Starlark type and expose it
The `PythonPackagingPolicy` dictates a lot of behavior for how
Python resources are packaged. Up to this point, our strategy
for managing these settings has been to pass arguments to
`PythonDistribution.to_python_executable()` that influence a
`PythonPackagingPolicy` created in the background. As complexity
increases, using arguments to influence policy creation is not
practical.

This commit introduces a Starlark wrapper type for representing
`PythonPackagingPolicy` instances. Our eventual plan is to feed this
into `to_python_executable()` instead of a myriad of arguments.

The new type currently does nothing at all except exist. This will
change in future commits.