py_packaging: pass PythonResourceAddCollectionContext to add_python_module_source
The PythonResourceAddCollectionContext type is much more expressive than
a simple location, as it allows us to influence more aspects of how a
resource is added to a collection.

This commit changes the binary builder API for adding a
PythonModuleSource to accept this type instead of a location.

We don't yet take full advantage of the new type. That will be added
in subsequent commits.