python-packaging: add method for adding module source with add context
This method uses the passed PythonResourceAddCollectionContext to
conditionally add the PythonModuleSource to the collection, possibly
performing multiple mutations, including deriving a bytecode request
as appropriate.

This method is more of a convenience method. But having it live in
the collector makes sense as the logic is non-trivial and the collector
module owns the PythonResourceAddCollectionContext type, so it makes
sense for the logic to live in this module.