starlark: use collection add context on Starlark Value
Previously, we were unconditionally deriving the context. Let's use
the context on the Starlark value if one is stored.

At this point, the add context modified by Starlark code should be
reflected all the way down to the low-level resource collection adder.
However, we do not yet expose the full functionality of the context
to Starlark. And we're not following this pattern for every Python
resource type. There's still a lot of work to be done.