python-packaging: add a method for adding a bytecode request with add context
We're slowly building out the APIs for adding resources with add
contexts.

This one is a bit wonky because the bytecode request type is associated
with a single optimization level. So we can only add a single
resource-location per call. So this method isn't as useful as other
ones. But API coverage seems useful to have.