starlark: expose more add context attributes
This allows Starlark to query and mutate all exposed fields of the add
collection context. This gives Starlark a lot more power over
influencing what happens when resources are added to binaries.

There's still a lot of room to add tests, limit attributes on a
per-type basis, and fully document the semantics. But this is a
start.