py_packaging: pass an add collection context to add_python_extension_module()
The logic in the function is a bit janky and the error messages no
longer reflect reality. But this was the path of least resistance to
getting this API to accept an add context. We'll overhaul this code
in future commits.

As part of this, we add the add context to the Starlark value. This
means all Starlark types representing Python resources now have add
contexts associated with them!