py_packaging: remove some soon to be redundant tests
In the new world, the policy doesn't matter at add time because we
can derive an add context if one isn't passed. So tests for an explicit
override aren't that interesting to us. What is interesting are the
combinations of `location` and `location_fallback`. We only support
the exact combinations derived from the policy. So let's only test
for those.