py_packaging: rewrite in terms of locations present in add context
This makes the logic similar to other Python resource types. Still not
exactly the same. But at least we're keying off the locations directly
and not querying the policy.

Because Starlark can now set a preferred location of the filesystem
and a fallback of memory (which isn't tested), we explicitly deny this
configuration.