starlark: remove location argument from add_* methods
All Python resource types now expose `add_*` attributes to control how
they are added and all methods now respect this. While we're still
missing functionality like the ability for Starlark to register a
callable to influence the default values, this mechanism is more
powerful and simpler than having both a default location and an
argument override. So the location form is being removed.