starlark: start to define callbacks for influencing add collection context
We want Starlark to be able to register callables that can influence the
`add_*` attributes on Python resources before they are added. This will
make it easier for Starlark code to bulk influence resource adding since
you will be able to define a single function influencing all resources.

We start this by teaching the PythonPackagingPolicy type to hold Values
of callables.