starlark: move policy application to PythonPackagingPolicyValue
This is logical best place for this functionality.

As part of this, we needed to track a Starlark value on the
`PythonExecutable` Starlark type to hold the Value.

And since the new method returns ValueResult, we had to change our
type coercion function to return ValueResult as well. This was a
long overdue change.