starlark: call resource callbacks when applying policy to resource
We recently taught Starlark to register functions that can influence
resource creation. This commit adds the calling of these functions as
part of applying the packaging policy to a resource.

This required threading TypeValues and CallStack through various
functions.

The mechanism by which we deal with Value/TypeValue here is a bit wonky
(see inline comment in python_packaging_policy.rs). But it gets the
job done. It also may conveniently help us avoid nested borrows, which
can cause runtime panics from Starlark.