py_packaging: add a builder trait for initializing from a distribution
It logically makes sense to separate population of the core interpreter
from other components, such as the Python standard library.

But our real reason for doing this is we want to provide a callback
function for influencing resource creation so custom Starlark code can
modify Python resources from the distribution. We could shoehorn this
into as_python_executable_builder(). But that function is already quite
large. I think it will be easier to maintain as a dedicated method on
the builder trait.