python-packaging: add support for reading Python wheels
I haven't tested this. But the wheel format is pretty trivial.

We want to implement support for reading wheel archives directly so that
we can perform Python packaging activities without having to invoke
`pip install`. This will enable us to perform (and test) packaging
on architectures that aren't the target architecture. While this will
eventually unblock cross-compiling, my main reason for doing this is
so we can start testing how non-native architecture wheels are handled
from any host architecture. e.g. we can write tests that run on Linux
that verify Windows wheels and binaries are handled correctly.

Wheels are zip files. And because the `zip` crate introduces a handful
of dependencies which we don't want to leak into the `pyembed` crate,
slowing down builds, we define some crate features to make the building
of wheel functionality optional.