py_packaging: initial support for `pip download`
This commit teaches the underlying packaging system and the binary
builder trait to invoke `pip download` to download wheels and emit
resources for what was downloaded. It uses the support for reading
wheel archives that was recently implemented.

A test is added demonstrating that we're capable of downloading
the `zstandard` wheel, which is a Python package I also maintain.

The new code is capable of using a different host distribution from
the target distribution. And the added test demonstrates that we're
capable of downloading wheels from non-native platforms and
extracting all relevant Python resources within. This will form
a nice foundation to support cross-compiling and will facilitate
cross-platform testing of wheels in the wild.