py_packaging: add a test verifying we can `pip download` numpy
numpy is a pretty popular package and it is doing some wonky things
with its wheel layout, notably installing support shared library
files in directories like `numpy.libs` and relying on the loader
relative path embedded in extension modules to find them.

Let's add test coverage for downloading numpy so that we can start
to implement support for it.