starlark: construct separate Python distribution for host if needed
Now that we have support for using different Python distributions for
the host and target triples, let's actually instantiate a host-specific
distribution from Starlark if the host triple doesn't match the build
triple.

This likely needs more logic, as there are cases where a target that
isn't the host can still run on the current machine. But we can deal
with this later.