python-packaging: track filename on libraries
Previously, we lost the library filename extension as part of
converting library dependencies to shared libraries and when adding
shared libraries as relative path files in a resource collector.

This commit teaches library dependencies to capture the original
filename and preserve it all the way through to the resource collector,
which will use the filename to materialize the library.

The test changes in the standalone builder now show that .dll
dependencies from distribution extension modules are properly
materialized on the filesystem. This fixes a long-standing bug
where we failed to properly install the .dll dependencies when
add extension modules to the filesystem, not memory.