python-packed-resources: create new resource fields to hold type flags
A single flavor field/enum to declare the resource type was not
sufficient because there are cases where a named resource could be
multiple types. For example, on Windows, `sqlite3` is both a Python
module and a shared library.

To handle this scenario, we can either store multiple maps of
named resources, segmenting by flavor. Or we can allow each named
resource to express multiple types. Our `Resource` type already
allows the expression of multiple types through different populated
fields. And maintainng multiple maps creates additional code
complexity. So we choose to go the route of allowing each resource
to express itself as multiple types.

Going forward, the `flavor` field is effectively deprecated. But
we still need to teach a lot of code to populate and read from the
new fields before we can truly make it optional.

Strictly speaking, the introduction of new fields is backwards
incompatible. We should consider bumping the resources format
version before release to express this.