pyembed: use resource type identification fields instead of flavor
This commits removes run-time logic from the importer that consults
the deprecated ResourceFlavor enum to determine how to proceed. This
should enable us to have resources that are both Python modules and
shared libraries.

We still need to teach the Python types in the crate about the new
resource fields.