py_packaging: expose and use python executable for building Rust
When we started to distinguish between host and target distributions,
the code for invoking cargo that passes PYTHON_SYS_EXECUTABLE was
inadvertently using the host triple's executable. When cross compiling
(and this includes between x86/x64 on Windows), we would get an
architecture mismatch when linking.

By using the host distribution's Python executable, we avoid this
problem.

This is an ugly workaround and it effectively goes one step backwards
on support for cross-compiling. But it should unbust x86 builds
from 64-bit binaries on Windows.