py_packaging: use target distribution on host if compatible
Previously, we did a blind compare of the host and target triples to
determine whether the target distribution could run on the host. In
reality, some distributions can run on multiple host targets. For
example, a 32-bit Windows distribution can run on 64-bit Windows. Or
a musl libc Linux distribution can run in a GNU Linux environment.

This commit teaches the distributions to emit what target triples
they are compatible with and for our logic of choosing a separate host
distribution to take this into account.