python-packaging: move logic for adding extension modules with context
Before this commit, the collector had all the logic for determining
how to add a resource given an add context except the logic for
extension modules, which is a bit complex.

Now that we have various types in the python-packaging crate, we can
move this logic into the crate/collector as well.

Some of this logic is somewhat PyOxidizer specific. Over time we'll
hopefully make it more generic. In the mean time, this reduces the
surface area of the builder code substantially.