python-packed-resources: formally introduce version 2 of the format
This should have been done when we introduced the new resource fields.
But I was being lazy when I wrote those commits. Let's do the version
bump now.

As part of this, I dropped support for reading/writing the old format
version. This could be preserved easily enough. But, agian, I'm lazy.