docs: remove PythonConfig struct
This type has effectively been deprecated since OxidizedPythonInterpreterConfig
was introduced. But I never spent the effort to remove it until now.

This includes a lot of docs updates that should have been done in the
previous commit. Whatever.

With the old config type out of the way, this makes some future
refactoring simpler, as there is only 1 type to worry about.