python-packaging: move PythonInterpreterConfig from pyembed
As part of moving all interpreter configuration types to a central
crate so they can be shared.

Since the type now lives in a separate crate and we don't want
to introduce a dependency on python3-sys, we had to refactor the
type conversion functions, as we can't implement a trait on a type
outside the current crate.