py_packaging: be more careful about adding builtin extension modules
Previously, if we were dynamically linking against libpython (read:
using an existing libpython instead of building our own) and we added
an extension module, it could be marked as a built-in. But the
extension module wouldn't actually have its object data linked into
the already existing libpython.

This commit teaches the extension module handling mechanism to be a
bit smarter.

Extension modules now advertise (perhaps not robustly enough) that they
are available in libpython. And resource collections now have a flag
indicating whether they allow new builtins that aren't already present.

The logic for adding a new extension module takes these into account:
we can only add a builtin extension module if it already exists as such
or if we're allowed to add new builtin extension modules.

The change in logic has a number of consequences for tests. Looking
at the test changes, I believe the new behavior to be reasonable.

This closes #249 and likely a handful of other issues.