python-packaging: store allowed locations on packaging policy
We want to get rid of the resources policy because it is an
avoidable abstraction. Its original purpose has been largely replaced
by the packaging policy.

This commit switches the packaging policy to hold a location and
fallback location instead of a resource policy and updates a lot of
code to reflect the new mechanism.