starlark: implement attributes for all config fields
Passing arguments to initialize PythonInterpreterConfig is a
lot of work and can be unwieldy. In future commits we'll be
transitioning to an attribute-based approach. And we'll initialize
instances from the Python distribution from which it came.

A prerequisite to this is exposing attributes for each inner field.
This commit implements that. We've held off on testing and documentation
for now because this commit is already quite large and unwieldy.