py_packaging: create a more reasonable default configuration
When we ported to PythonInterpreterConfig, we carried over setting
values to Some(T) instead of relying on the Python interpreter
defaults initialized from the profile.

I believe it is reasonably safe to set these to None by default and
let pyembed deal with any initialization, if needed. If this breaks
things, we should have had test code to catch regressions!