py_packaging: only emit single PythonExtensionModule if we built it
Previously, we could emit a PythonExtensionModule corresponding to
the shared library found during filesystem scanning and a subsequent
one built from the metadata collected by our hacked distutils.

If we were using the hacked distutils, that is the extension module
we should be using. Also, this PythonExtensionModule is likely a
superset of the original and should generally be the better version
to use.

This _might_ be enough to address #293. But there are still gaps in
library detection preventing me from calling this a highly robust
solution. It does remove a kludge and should enable us to link built
extensions as built-ins again.