pyembed: store explicit argv value on OxidizedPythonInterpreterConfig
Our sys.argv initialization is buggy (#294). In preparation for fixing
it, we teach OxidizedPythonInterpreterConfig to store an optional vector
of OsString to initialize it with. If not defined and not overridden
by the low-level interpreter config, we will use `std::env::args_os()`
like we do currently.

Having an explicit field to control `sys.argv` initialization enables
us to write unit tests around value handling, which we've done.

The tests expose some wonky behavior, such as us not chomping off
the first value. (`sys.argv` doesn't contain the executable program
name as the first element like `int main()` provides.)