project_building: use duct to run cargo
This enables us to redirect stderr to stdout so all build output
is captured and sent to the logger.

A benefit of this change is that tests no longer spew output to
stderr, polluting output. A downside is that because stdout/stderr
are now pipes instead of a TTY, cargo doesn't colorize output. I
believe there are environment variables to force terminal output
mode and we could potentially enable them if the original stdin is
a TTY. I've left this as a TODO.