starlark: store and use a DistributionCache on EnvironmentContext
This will enable us to leverage a global cache in Starlark tests.

We don't yet use it in Starlark tests. I wanted to isolate the
introduction of this change from actually using it so we could
isolate any changes.

I don't believe this change impacts test timing at all: both wall
and CPU time are unchanged.