starlark: use distribution cache in tests
We supported it in the previous commits. Let's use it in
tests.

On my Ryzen 3950X, this drops test execution times accordingly in
debug builds:

before: wall: 97s; cpu 750s
after:  wall: 97s; cpu 352s

On release builds, wall time is unchanged and overall CPU time drops
from ~230s to ~190s. I guess the debug code for scanning and loading
Python distributions is just really slow!

So we dropped overall CPU utilization of tests to ~47% of original.
Nice!

This doesn't matter as much for me as it does for environments
with fewer CPU cores, like our CI environment, which is really
struggling with job times at the moment. Hopefully this speeds
things up!