pyoxidizer: embed an application manifest enabling long paths support
Before, long paths support was highly subject to the build or run-time
environment. By embedding an application manifest, we bolster the
chances of long paths working.

I'm unable to verify this since modern versions of Windows 10 enable
long paths by default. But the resource including the manifest is
being embedded in the library via dynamic linking via build.rs. Since
the resource is embedded, I'm pretty sure this "just works."

I imagine someday someone will request a feature to manipulate the
content of the application manifest and we'll want to introduce Starlark
functionality to facilitate that. Until that day arives, let's avoid
the complexity.

Closes #197.