pyembed: set missing path config fields in all modes by default
I think this default makes sense because the path layout
compiled into libpython will likely not match the run-time layout,
at least in the context of PyOxidizer. From the context of an
executable embedding Python, you likely control the embedded
Python interpreter install and you want to explicitly set its
path config based on the currently running executable.

There is potentially room to have this set to false by default
in pyembed, as the Python picked up by the build script should
theoretically have a reasonable layout. But it definitely needs
to be enabled by PyOxidizer.