python-packed-resources: allow arbitrary file storage, bump version to 3
Our approach of classifying a resource as a strongly typed resource
is useful for run-time performance. But it requires us to properly
classify every file at packaging time. This has proved to be a
difficult endeavor and getting every corner case correct is going to
take a while.

While I feel strongly that strongly typed resources is the better
approach, every day that passes where the approach isn't compatible
with a package in the wild is a day where a potential user can't use
PyOxidizer. Rather than force those people to wait, I'm carving out
support in the packed resources data structure for encoding arbitrary
filenames.

The way this will work is if the "is filename" bit is set, the resource
name is actually a UTF-8 encoded filename. That file's data is either
embedded in the data structure or referenced via a relative filesystem
path.

The goal here is that we collect and index a collection of filenames
and then materialize those filenames at run-time and fall back to
Python's built-in, filesystem based importers. We can teach pyembed to
extract files to a temporary directory, tmpfs, squashfs, whatever, and
PyOxidizer should have parity with other Python packaging tools that
take the filesystem-based approach to resource management.