pyoxidizer: support installing tcl support files
Previous commits taught distributions to expose tcl files and for
the run-time code to define TCL_LIBRARY.

This commit hooks it all up and teaches executable builders to set
a directory to install tcl files into. When used, we install extra
files next to the build executable and configure the built executable
to load tcl files from the installed location.

I've tested this on Linux and Windows and it seems to _just work_.

Closes #25.