pyoxidizer: remove the concept of a run mode
It turns out that we had refactored the internals to populate
the PyConfig fields based on this value. So this feature was redundant
with the fields in PyConfig.

That redundancy was creating complexity and possibly confusion. I think
the best thing to do is remove the feature and keep things simpler.

This commit removes the `run_mode` field from the run-time config.
As we pulled on that thread, it turns out nothing uses the
`PythonRunMode` enum any more. So it has been removed.

We could retain support in Starlark for translating old attribute
sets to the inner config sets. But I'm inclined to take the BC break
since PyOxidizer isn't yet stable.