docs: move docs for oxidized_importer closer to their source
The repository logically consists of multiple projects. But the
docs/ directory has documentation for all of them. I want to make
it possible for each project to have its own documentation.

This commit starts that effort by establishing a pyembed/docs
directory and moving the oxidized_importer docs there. Arguably
oxidized_importer should have its own docs. But the code is in
the pyembed crate (for now) and I think it makes sense for the
docs to live there instead of in the oxidized_importer directory.

We also enable warnings as errors when running `sphinx-build` and
add CI jobs to build the docs.

The copying of the .rst files is a bit hacky. But it is low effort.